finished removing deprecated billing app
